refactor(channels): source channel health checks from ChannelRegistry
