+++ app/Channels/Linear/HealthCheck.php
@@ -4,7 +4,7 @@ namespace App\Channels\Linear;
 
 use App\Exceptions\LinearOAuthRefreshFailedException;
 use App\Models\LinearOauthConnection;
-use App\Services\HealthCheck\Channel\ChannelCheck;
+use App\Services\HealthCheck\ChannelCheck;
 use App\Services\HealthCheck\HealthAction;
 use App\Services\HealthCheck\HealthResult;
 use Illuminate\Support\Facades\Http;
